Chore: Refactor done
