+++ lib/screen/home_screen/imports.dart
@@ -1,0 +1,7 @@
+export 'package:flutter/material.dart';
+export '../../widgets/bottom_nav_bar.dart';
+export '../widgets/banner_card.dart';
+export '../widgets/categories_list.dart';
+export '../widgets/home_appbar.dart';
+export '../widgets/recommendation.dart';
+export '../widgets/title_heading.dart';
