+++ lib/screen/widgets/categories_list.dart
@@ -1,0 +1,52 @@
+import 'package:cached_network_image/cached_network_image.dart';
+import 'package:flutter/material.dart';
+
+import '../../themes/color_theme.dart';
+
+class CategoriesList extends StatelessWidget {
+  const CategoriesList({
+    super.key,
+    required this.images,
+  });
+
+  final List images;
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 30, top: 15),
+      child: SingleChildScrollView(
+        scrollDirection: Axis.horizontal,
+        child: Row(
+          children: List.generate(
+              12,
+              (index) => Padding(
+                    padding: const EdgeInsets.only(right: 20),
+                    child: Column(
+                      children: [
+                        CircleAvatar(
+                          radius: 30,
+                          backgroundColor: AppColor.darkBackgroundColor,
+                          child: ClipRRect(
+                            borderRadius: BorderRadius.circular(40),
+                            child: CachedNetworkImage(
+                              fit: BoxFit.fill,
+                              imageUrl: images[index],
+                              placeholder: (context, url) =>
+                                  const CircularProgressIndicator(),
+                              errorWidget: (context, url, error) =>
+                                  const Icon(Icons.error),
+                            ),
+                          ),
+                        ),
+                        Padding(
+                            padding: const EdgeInsets.only(top: 10),
+                            child: Text('$index ghb nn'))
+                      ],
+                    ),
+                  )),
+        ),
+      ),
+    );
+  }
+}
